Use Pylast to get artists data.
Nowadays LastFM requires user to be logged in to fetch all library data.
Short adventures with login & fetch didn't succeed so this is an
alternative fix for the problem. Drawback is that it's dead slow
compared to previous artist retrieving. I'll might revisit this if I
have time and interest.

Signed-off-by: Ville Valkonen <weezel@users.noreply.github.com>